+++ src/sas/qtgui/Perspectives/Corfunc/Slider.py
