+++ src/sas/qtgui/Perspectives/Corfunc/Util.py
